+++ api/utils.py
@@ -84,11 +84,11 @@ class APIUtils():
 
         for key, value in aggregations.items():
             if key == 'buckets':
+                print(key, value)
                 aggs[parentKey] = [
-                    {
-                        'value': b['key'],
-                        'count': b['editions_per']['doc_count']
-                    }
+                    {'value': b['key'], 'count': b['editions_per']['doc_count']}
+                    if 'editions_per' in b.keys() else 
+                    {'value': b['key_as_string'],'count': b['doc_count']}
                     for b in value
                 ]
                 return aggs